+++ cmd/thanos/compact.go
@@ -9,12 +9,9 @@ import (
 	"github.com/go-kit/kit/log/level"
 	"github.com/improbable-eng/thanos/pkg/compact"
 	"github.com/improbable-eng/thanos/pkg/compact/downsample"
-	"github.com/improbable-eng/thanos/pkg/objstore/azure"
 	"github.com/improbable-eng/thanos/pkg/objstore/client"
-	"github.com/improbable-eng/thanos/pkg/objstore/s3"
 	"github.com/improbable-eng/thanos/pkg/runutil"
 	"github.com/oklog/run"
-	"github.com/oklog/ulid"
 	"github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
@@ -33,15 +30,13 @@ func registerCompact(m map[string]setupFunc, app *kingpin.Application, name stri
 	dataDir := cmd.Flag("data-dir", "Data directory in which to cache blocks and process compactions.").
 		Default("./data").String()
 
-	gcsBucket := cmd.Flag("gcs.bucket", "Google Cloud Storage bucket name for stored blocks.").
-		PlaceHolder("<bucket>").String()
+	bucketConf := cmd.Flag("objstore.config", "The object store configuration in yaml format.").
+		PlaceHolder("<bucket.config.yaml>").Required().String()
 
-	s3config := s3.RegisterS3Params(cmd)
+	syncDelay := modelDuration(cmd.Flag("sync-delay", "Minimum age of fresh (non-compacted) blocks before they are being processed.").
+		Default("30m"))
 
-	azureConfig := azure.RegisterAzureParams(cmd)
-
-	syncDelay := cmd.Flag("sync-delay", "Minimum age of fresh (non-compacted) blocks before they are being processed.").
-		Default("30m").Duration()
+	retention := modelDuration(cmd.Flag("retention.default", "How long to retain samples in bucket. 0d - disables retention").Default("0d"))
 
 	wait := cmd.Flag("wait", "Do not exit after all compactions have been processed and wait for new work.").
 		Short('w').Bool()
@@ -50,12 +45,11 @@ func registerCompact(m map[string]setupFunc, app *kingpin.Application, name stri
 		return runCompact(g, logger, reg,
 			*httpAddr,
 			*dataDir,
-			*gcsBucket,
-			s3config,
-			azureConfig,
-			*syncDelay,
+			*bucketConf,
+			time.Duration(*syncDelay),
 			*haltOnError,
 			*wait,
+			time.Duration(*retention),
 			name,
 		)
 	}
@@ -67,12 +61,11 @@ func runCompact(
 	reg *prometheus.Registry,
 	httpBindAddr string,
 	dataDir string,
-	gcsBucket string,
-	s3Config *s3.Config,
-	azureConfig *azure.Config,
+	bucketConf string,
 	syncDelay time.Duration,
 	haltOnError bool,
 	wait bool,
+	retention time.Duration,
 	component string,
 ) error {
 	halted := prometheus.NewGauge(prometheus.GaugeOpts{
@@ -87,7 +80,7 @@ func runCompact(
 
 	reg.MustRegister(halted)
 
-	bkt, err := client.NewBucket(logger, &gcsBucket, *s3Config, *azureConfig, reg, component)
+	bkt, err := client.NewBucket(logger, bucketConf, reg, component)
 	if err != nil {
 		return err
 	}
@@ -101,134 +94,104 @@ func runCompact(
 
 	sy, err := compact.NewSyncer(logger, reg, bkt, syncDelay)
 	if err != nil {
-		return err
+		return errors.Wrap(err, "create syncer")
 	}
-	// Start cycle of syncing blocks from the bucket and garbage collecting the bucket.
-	{
-		// Instantiate the compactor with different time slices. Timestamps in TSDB
-		// are in milliseconds.
-		comp, err := tsdb.NewLeveledCompactor(reg, logger, []int64{
-			int64(1 * time.Hour / time.Millisecond),
-			int64(2 * time.Hour / time.Millisecond),
-			int64(8 * time.Hour / time.Millisecond),
-			int64(2 * 24 * time.Hour / time.Millisecond),  // 2 days
-			int64(14 * 24 * time.Hour / time.Millisecond), // 2 weeks
-		}, downsample.NewPool())
-		if err != nil {
-			return errors.Wrap(err, "create compactor")
+
+	// Instantiate the compactor with different time slices. Timestamps in TSDB
+	// are in milliseconds.
+	comp, err := tsdb.NewLeveledCompactor(reg, logger, []int64{
+		int64(1 * time.Hour / time.Millisecond),
+		int64(2 * time.Hour / time.Millisecond),
+		int64(8 * time.Hour / time.Millisecond),
+		int64(2 * 24 * time.Hour / time.Millisecond),  // 2 days
+		int64(14 * 24 * time.Hour / time.Millisecond), // 2 weeks
+	}, downsample.NewPool())
+	if err != nil {
+		return errors.Wrap(err, "create compactor")
+	}
+
+	var (
+		compactDir      = path.Join(dataDir, "compact")
+		downsamplingDir = path.Join(dataDir, "downsample")
+	)
+
+	compactor := compact.NewBucketCompactor(logger, sy, comp, compactDir, bkt)
+
+	if retention.Seconds() != 0 {
+		level.Info(logger).Log("msg", "default retention policy is enabled", "duration", retention)
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	f := func() error {
+		if err := compactor.Compact(ctx); err != nil {
+			return errors.Wrap(err, "compaction failed")
 		}
 
-		ctx, cancel := context.WithCancel(context.Background())
+		// After all compactions are done, work down the downsampling backlog.
+		// We run two passes of this to ensure that the 1h downsampling is generated
+		// for 5m downsamplings created in the first run.
+		level.Info(logger).Log("msg", "start first pass of downsampling")
 
-		f := func() error {
-			var (
-				compactDir      = path.Join(dataDir, "compact")
-				downsamplingDir = path.Join(dataDir, "downsample")
-			)
-
-			// Loop over bucket and compact until there's no work left.
-			for {
-				level.Info(logger).Log("msg", "start sync of metas")
-
-				if err := sy.SyncMetas(ctx); err != nil {
-					return errors.Wrap(err, "sync")
-				}
-
-				level.Info(logger).Log("msg", "start of GC")
-
-				if err := sy.GarbageCollect(ctx); err != nil {
-					return errors.Wrap(err, "garbage")
-				}
-
-				groups, err := sy.Groups()
-				if err != nil {
-					return errors.Wrap(err, "build compaction groups")
-				}
-				done := true
-				for _, g := range groups {
-					id, err := g.Compact(ctx, compactDir, comp)
-					if err == nil {
-						// If the returned ID has a zero value, the group had no blocks to be compacted.
-						// We keep going through the outer loop until no group has any work left.
-						if id != (ulid.ULID{}) {
-							done = false
-						}
-						continue
-					}
-
-					if compact.IsIssue347Error(err) {
-						err = compact.RepairIssue347(ctx, logger, bkt, err)
-						if err == nil {
-							done = false
-							continue
-						}
-					}
-					return errors.Wrap(err, "compaction")
-				}
-				if done {
-					break
-				}
-			}
-
-			// After all compactions are done, work down the downsampling backlog.
-			// We run two passes of this to ensure that the 1h downsampling is generated
-			// for 5m downsamplings created in the first run.
-			level.Info(logger).Log("msg", "start first pass of downsampling")
-
-			if err := downsampleBucket(ctx, logger, bkt, downsamplingDir); err != nil {
-				return errors.Wrap(err, "first pass of downsampling failed")
-			}
-
-			level.Info(logger).Log("msg", "start second pass of downsampling")
-
-			if err := downsampleBucket(ctx, logger, bkt, downsamplingDir); err != nil {
-				return errors.Wrap(err, "second pass of downsampling failed")
-			}
-
-			level.Info(logger).Log("msg", "compaction iteration done")
-			return nil
+		if err := downsampleBucket(ctx, logger, bkt, downsamplingDir); err != nil {
+			return errors.Wrap(err, "first pass of downsampling failed")
 		}
 
-		g.Add(func() error {
-			defer runutil.CloseWithLogOnErr(logger, bkt, "bucket client")
+		level.Info(logger).Log("msg", "start second pass of downsampling")
 
-			if !wait {
-				return f()
+		if err := downsampleBucket(ctx, logger, bkt, downsamplingDir); err != nil {
+			return errors.Wrap(err, "second pass of downsampling failed")
+		}
+
+		if retention.Seconds() != 0 {
+			if err := compact.ApplyDefaultRetentionPolicy(ctx, logger, bkt, retention); err != nil {
+				return errors.Wrap(err, "retention failed")
+			}
+		}
+		level.Info(logger).Log("msg", "compaction, downsampling and optional retention apply iteration done")
+		return nil
+	}
+
+	g.Add(func() error {
+		defer runutil.CloseWithLogOnErr(logger, bkt, "bucket client")
+
+		if !wait {
+			return f()
+		}
+
+		// --wait=true is specified.
+		return runutil.Repeat(5*time.Minute, ctx.Done(), func() error {
+			err := f()
+			if err == nil {
+				return nil
+			}
+			// The HaltError type signals that we hit a critical bug and should block
+			// for investigation.
+			// You should alert on this being halted.
+			if compact.IsHaltError(err) {
+				if haltOnError {
+					level.Error(logger).Log("msg", "critical error detected; halting", "err", err)
+					halted.Set(1)
+					select {}
+				} else {
+					return errors.Wrap(err, "critical error detected")
+				}
 			}
 
-			// --wait=true is specified.
-			return runutil.Repeat(5*time.Minute, ctx.Done(), func() error {
-				err := f()
-				if err != nil {
-					// The HaltError type signals that we hit a critical bug and should block
-					// for investigation.
-					// You should alert on this being halted.
-					if compact.IsHaltError(err) {
-						if haltOnError {
-							level.Error(logger).Log("msg", "critical error detected; halting", "err", err)
-							halted.Set(1)
-							select {}
-						} else {
-							return errors.Wrap(err, "critical error detected")
-						}
-					}
+			// The RetryError signals that we hit an retriable error (transient error, no connection).
+			// You should alert on this being triggered to frequently.
+			if compact.IsRetryError(err) {
+				level.Error(logger).Log("msg", "retriable error", "err", err)
+				retried.Inc()
+				// TODO(bplotka): use actual "retry()" here instead of waiting 5 minutes?
+				return nil
+			}
 
-					// The RetryError signals that we hit an retriable error (transient error, no connection).
-					// You should alert on this being triggered to frequently.
-					if compact.IsRetryError(err) {
-						level.Error(logger).Log("msg", "retriable error", "err", err)
-						retried.Inc()
-						// TODO(bplotka): use actual "retry()" here instead of waiting 5 minutes?
-						return nil
-					}
-				}
-
-				return err
-			})
-		}, func(error) {
-			cancel()
+			return errors.Wrap(err, "error executing compaction")
 		})
-	}
+	}, func(error) {
+		cancel()
+	})
+
 	if err := metricHTTPListenGroup(g, logger, reg, httpBindAddr); err != nil {
 		return err
 	}